==> scripts/3.data_sampling.ipynb <==
import pandas as pd
import dask.dataframe as dd
from dask_ml.model_selection import train_test_split
from dask.distributed import Client
from sklearn.model_selection import StratifiedShuffleSplit

df_pth = "N:/Classes/2023_1SPRING/Data 7900 - Prof. Franks - Applied Analysis Project/Data/subset/final_validation_df_v2.csv"

fold_pth = "N:/Classes/2023_1SPRING/Data 7900 - Prof. Franks - Applied Analysis Project/Data/subset/"



chunksize = 10_000_000
rand_state = 42023
n_splits = 2
test_size = 5_000_000

full_email = {"HASH_NBR": "int64", "CAMPAIGN_NBR": "int64", "UNSUB_IND": "int8", "SEND_DT": "object", "CLICK": "int8", "MBR_PRGM_ACTV": "int8", "week_of_year": "int16", "day_of_month": "int16", "day_of_year": "int16", "is_end_month": "int8", "is_start_month": "int8", "is_end_quarter": "int8", "is_start_quarter": "int8", "CC_2": "int8", "CC_3": "int8", "CC_4": "int8", "TIER_2": "int8", "TIER_3": "int8", "TIER_4": "int8", "quarter_1": "int8", "quarter_3": "int8", "quarter_4": "int8", "GNDR_1": "int8", "GNDR_2": "int8", "EN_1": "int8", "EN_2": "int8", "EN_3": "int8", "EN_4": "int8", "EN_6": "int8", "AGE_1": "int8", "AGE_2": "int8", "AGE_3": "int8", "AGE_4": "int8", "AGE_5": "int8", "INCOME_1": "int8", "INCOME_2": "int8", "INCOME_3": "int8", "INCOME_4": "int8", "INCOME_5": "int8", "January": "int8", "February": "int8", "April": "int8", "May": "int8", "June": "int8", "July": "int8", "August": "int8", "September": "int8", "October": "int8", "November": "int8", "December": "int8", "Monday": "int8", "Tuesday": "int8", "Wednesday": "int8", "Thursday": "int8", "Friday": "int8", "Saturday": "int8", "Fall": "int8", "Summer": "int8", "Winter": "int8", "Oceania": "int8", "Asia_Africa": "int8", "AUSTRALASIA_JAPAN": "int8", "CANADA": "int8", "CARIBBEAN": "int8", "EUROPE": "int8", "HONG_KONG": "int8", "INDIA_MIDDLE_EAST_AFRICA": "int8", "LATIN_AMERICA": "int8", "MAINLAND_CHINA": "int8", "MEXICO": "int8", "SOUTHEAST_ASIA_KOREA": "int8", "TAIWAN": "int8", "MACAU": "int8", "CONF_HASH_NBR": "int64", "HTL_HASH_NBR": "int64", "NUM_UNIQUE_CONF_HASH_NBR": "int16", "NUM_UNIQUE_HTL_HASH_NBR": "int16", "BUS_LEIS_IND": "int8", "STAY_REVENUE_USD": "float64", "TIME_BWTN_CONF_CKIN_DT": "int16", "TIME_BWTN_CKIN_CKOUT_DT": "int16", "CONF_DT_daysto_holiday": "int16", "CK_IN_DT_daysto_holiday": "int16", "CK_OUT_DT_daysto_holiday": "int16", "CONF_DT_nearest_holiday_dayofyear": "int16", "CK_IN_DT_nearest_holiday_dayofyear": "int16", "CK_OUT_DT_nearest_holiday_dayofyear": "int16", "DIFFERENT_CONTINENT": "int8", "MBR_TENURE_ASOF_CONF_DT": "int16", "MBR_TENURE_ASOF_CK_IN_DT": "int16", "MBR_TENURE_ASOF_CK_OUT_DT": "int16", "HTL_RGN_EMEAA": "int8", "HTL_RGN_GCHINA": "int8", "HTL_CHAIN_CATEGORY_CHN_CAT_1": "int8", "HTL_CHAIN_CATEGORY_CHN_CAT_2": "int8", "HTL_CHAIN_CATEGORY_CHN_CAT_3": "int8", "REWARD_NT_RN_1": "int8", "REWARD_NT_RN_2": "int8", "HTL_CHAIN_CHN_2": "int8", "HTL_CHAIN_CHN_3": "int8", "HTL_CHAIN_CHN_4": "int8", "HTL_CHAIN_CHN_5": "int8", "HTL_CHAIN_CHN_6": "int8", "HTL_CHAIN_CHN_7": "int8", "HTL_CHAIN_CHN_9": "int8", "HTL_CHAIN_CHN_10": "int8", "HTL_CHAIN_CHN_11": "int8", "HTL_CHAIN_CHN_12": "int8", "HTL_CHAIN_CHN_13": "int8", "HTL_CHAIN_CHN_14": "int8", "HTL_CHAIN_CHN_15": "int8", "HTL_CHAIN_CHN_18": "int8", "HTL_CHAIN_CHN_20": "int8", "HTL_CHAIN_CHN_22": "int8", "HTL_CHAIN_CHN_23": "int8", "NUM_STAYS": "int16", "TOTAL_STAY_REVENUE": "float64", "TOTAL_ROOMS": "int16", "TOTAL_GUEST_QTY": "int16", "TOTAL_TIME_BWTN_CONF_CKIN_DT": "int16", "TOTAL_TIME_BWTN_CKIN_CKOUT_DT": "int16", "TOTAL_CONF_DT_daysto_holiday": "int16", "TOTAL_CK_IN_DT_daysto_holiday": "int16", "TOTAL_CK_OUT_DT_daysto_holiday": "int16", "TOTAL_CONF_DT_nearest_holiday_dayofyear": "int16", "TOTAL_CK_IN_DT_nearest_holiday_dayofyear": "int16", "TOTAL_CK_OUT_DT_nearest_holiday_dayofyear": "int16", "NUM_BUS_LEIS_IND": "int16", "TOTAL_BUS_LEIS_IND_REVENUE": "float64", "TOTAL_BUS_LEIS_IND_ROOMS": "int16", "TOTAL_BUS_LEIS_IND_GUEST_QTY": "int16", "NUM_DIFFERENT_CONTINENT": "int16", "TOTAL_DIFFERENT_CONTINENT_REVENUE": "float64", "TOTAL_DIFFERENT_CONTINENT_ROOMS": "int16", "TOTAL_DIFFERENT_CONTINENT_GUEST_QTY": "int16", "NUM_HTL_RGN_EMEAA": "int16", "TOTAL_HTL_RGN_EMEAA_REVENUE": "float64", "TOTAL_HTL_RGN_EMEAA_ROOMS": "int16", "TOTAL_HTL_RGN_EMEAA_GUEST_QTY": "int16", "NUM_HTL_RGN_GCHINA": "int16", "TOTAL_HTL_RGN_GCHINA_REVENUE": "float64", "TOTAL_HTL_RGN_GCHINA_ROOMS": "int16", "TOTAL_HTL_RGN_GCHINA_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CATEGORY_CHN_CAT_1": "int16", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_1_REVENUE": "float64", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_1_ROOMS": "int16", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_1_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CATEGORY_CHN_CAT_2": "int16", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_2_REVENUE": "float64", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_2_ROOMS": "int16", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_2_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CATEGORY_CHN_CAT_3": "int16", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_3_REVENUE": "float64", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_3_ROOMS": "int16", "TOTAL_HTL_CHAIN_CATEGORY_CHN_CAT_3_GUEST_QTY": "int16", "NUM_REWARD_NT_RN_1": "int16", "TOTAL_REWARD_NT_RN_1_REVENUE": "float64", "TOTAL_REWARD_NT_RN_1_ROOMS": "int16", "TOTAL_REWARD_NT_RN_1_GUEST_QTY": "int16", "NUM_REWARD_NT_RN_2": "int16", "TOTAL_REWARD_NT_RN_2_REVENUE": "float64", "TOTAL_REWARD_NT_RN_2_ROOMS": "int16", "TOTAL_REWARD_NT_RN_2_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_2": "int16", "TOTAL_HTL_CHAIN_CHN_2_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_2_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_2_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_3": "int16", "TOTAL_HTL_CHAIN_CHN_3_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_3_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_3_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_4": "int16", "TOTAL_HTL_CHAIN_CHN_4_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_4_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_4_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_5": "int16", "TOTAL_HTL_CHAIN_CHN_5_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_5_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_5_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_6": "int16", "TOTAL_HTL_CHAIN_CHN_6_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_6_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_6_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_7": "int16", "TOTAL_HTL_CHAIN_CHN_7_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_7_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_7_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_9": "int16", "TOTAL_HTL_CHAIN_CHN_9_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_9_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_9_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_10": "int16", "TOTAL_HTL_CHAIN_CHN_10_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_10_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_10_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_11": "int16", "TOTAL_HTL_CHAIN_CHN_11_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_11_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_11_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_12": "int16", "TOTAL_HTL_CHAIN_CHN_12_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_12_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_12_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_13": "int16", "TOTAL_HTL_CHAIN_CHN_13_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_13_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_13_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_14": "int16", "TOTAL_HTL_CHAIN_CHN_14_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_14_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_14_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_15": "int16", "TOTAL_HTL_CHAIN_CHN_15_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_15_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_15_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_18": "int16", "TOTAL_HTL_CHAIN_CHN_18_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_18_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_18_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_20": "int16", "TOTAL_HTL_CHAIN_CHN_20_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_20_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_20_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_22": "int16", "TOTAL_HTL_CHAIN_CHN_22_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_22_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_22_GUEST_QTY": "int16", "NUM_HTL_CHAIN_CHN_23": "int16", "TOTAL_HTL_CHAIN_CHN_23_REVENUE": "float64", "TOTAL_HTL_CHAIN_CHN_23_ROOMS": "int16", "TOTAL_HTL_CHAIN_CHN_23_GUEST_QTY": "int16", "MBR_TENURE": "int16"}

col_nms = pd.read_csv(df_pth, nrows = 1).columns.to_list()


%%time

df_chunks = pd.read_csv(df_pth, 
                        skiprows = 1,
                        names = col_nms,
                        chunksize = chunksize,
                        parse_dates = ["SEND_DT"],
                        dtype = full_email
                       )

nr = 0
stratified_sample_list = []
validation_list = []
click_list = []
for i, df in enumerate(df_chunks):
    nr = nr + df.shape[0]
    stratified_split = StratifiedShuffleSplit(n_splits = 1, 
                                              test_size = 3612075, 
                                              random_state = rand_state
                                             )
                                            
    for train_index, sample_index in stratified_split.split(df, df["CLICK"]):
        stratified_sample = df.iloc[sample_index]
        validation_sample = df.iloc[train_index]

    print(f"{i} - {stratified_sample.CLICK.mean()} - {validation_sample.CLICK.mean()} - {df.CLICK.mean()}")

    stratified_sample_list.append(stratified_sample)
    validation_list.append(validation_sample)
    click_list = click_list + df.CLICK.tolist()

################################################################################################################

df = pd.concat(validation_list, axis = 0)

df.loc[df["CLICK"] == 1, "total_clicks"] = df.loc[df["CLICK"] == 1, "total_clicks"] - 1
df.loc[df["total_clicks"] < 0, "total_clicks"] = 0
df["CLICK_THROUGH_RATE"] = (df["total_clicks"] / df["total_emails"]).fillna(0)

df.to_csv(fold_pth + "final_validation_subdf.csv", index = False)

################################################################################################################

df = pd.concat(stratified_sample_list, axis = 0)

df.loc[df["CLICK"] == 1, "total_clicks"] = df.loc[df["CLICK"] == 1, "total_clicks"] - 1
df.loc[df["total_clicks"] < 0, "total_clicks"] = 0
df["CLICK_THROUGH_RATE"] = (df["total_clicks"] / df["total_emails"]).fillna(0)

df.to_csv(fold_pth + "final_email_subdf.csv", index = False)

################################################################################################################

print(df.shape[0])
df.head(3)

%%time

import datetime
import numpy as np
import gc

for k, ts in [(4, 1806038), (5, 903019), (6, 903019), (7, 722415)]: # 
    print(f"STARTING FILE: final_email_subdf_{k}.csv -- {datetime.datetime.now()}")
    df_chunks = pd.read_csv(fold_pth + f"final_email_subdf_{k}.csv", 
                            skiprows = 1,
                            names = col_nms,
                            chunksize = chunksize,
                            parse_dates = ["SEND_DT"],
                            dtype = full_email
                        )

    nr = 0
    stratified_sample_list = []
    click_list = []
    for i, df in enumerate(df_chunks):
        nr = nr + df.shape[0]
        stratified_split = StratifiedShuffleSplit(n_splits = 1, 
                                                  test_size = ts, 
                                                  random_state = rand_state
                                                 )
                                                
        for train_index, sample_index in stratified_split.split(df, df["CLICK"]):
            stratified_sample = df.iloc[sample_index]

        print(f"{i} - {stratified_sample.CLICK.mean()} - {df.CLICK.mean()}")

        stratified_sample_list.append(stratified_sample)
        click_list = click_list + df.CLICK.tolist()

    print(f"OVERALL CLICK RATE: {np.mean(click_list)}")

    ################################################################################################################

    print(f"SAVING TRAINING SET -- {datetime.datetime.now()}")

    df = pd.concat(stratified_sample_list, axis = 0)

    df.loc[df["CLICK"] == 1, "total_clicks"] = df.loc[df["CLICK"] == 1, "total_clicks"] - 1
    df.loc[df["total_clicks"] < 0, "total_clicks"] = 0
    df["CLICK_THROUGH_RATE"] = (df["total_clicks"] / df["total_emails"]).fillna(0)

    print(f"OVERALL TRAINING CLICK RATE: {df.CLICK.mean()}")

    df.to_csv(fold_pth + "final_email_subdf.csv", index = False, mode = 'a', header = False)

    print(f"SHAPE: {df.shape[0]}")

    del stratified_sample_list, click_list, nr, df, 
    gc.collect()

{'final_email_subdf_4.csv': 15907815,
 'final_email_subdf_5.csv': 34218924,
 'final_email_subdf_6.csv': 36204782,
 'final_email_subdf_7.csv': 42367075}
